begin refactoring event job mixins
